refactor(sip): extract BLF presence into the BlfSubscriptions state machine
Fourth extraction. BlfSubscriptions owns the fixed watcher-dialog slots
and the whole RFC 6665/4235 flow: event-package gate, AOR validation,
slot alloc/refresh/unsubscribe, 202 Accepted + immediate NOTIFY,
change-detection NOTIFYs and expiry sweep. Dialog state is computed
through a read-only sessionsView() on PbxEnv, which also grows
validAor() and requestedExpires() forwards.
